refactor: Migrate to src folder
