+++ src/ExtractedModule.js
