+++ src/schema/plugin.json
