+++ test/cases/optimize-tree/webpack.config.js
@@ -1,4 +1,4 @@
-const ExtractTextPlugin = require('../../../');
+const ExtractTextPlugin = require('../../../src/');
 const webpack = require('webpack');
 
 module.exports = {
