+++ test/cases/simple-query-object/webpack.config.js
@@ -1,4 +1,4 @@
-const ExtractTextPlugin = require('../../../');
+const ExtractTextPlugin = require('../../../src/');
 module.exports = {
   entry: './index',
   module: {
