+++ test/schema-validation.test.js
@@ -1,8 +1,8 @@
 // import ExtractTextPlugin from '../src';
 // const loader = require.resolve('../src/loader');
 /* eslint-disable */
-var ExtractTextPlugin = require("../");
-var loader = require.resolve('../loader');
+var ExtractTextPlugin = require("../src/");
+var loader = require.resolve('../src/loader');
 
 describe('json schema validation', () => {
   it('does not throw if a filename is specified', () => {
